Ensure restriction paths are ancestors of items when lowering to HIR
